code simplification, split of the main
