+++ src/main/java/fr/insalyon/creatis/moteurlite/MoteurLite.java
@@ -1,168 +1,19 @@
 package fr.insalyon.creatis.moteurlite;
 
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import org.apache.log4j.Logger;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-
-import fr.insalyon.creatis.gasw.Gasw;
-import fr.insalyon.creatis.gasw.GaswException;
-import fr.insalyon.creatis.gasw.GaswInput;
-import fr.insalyon.creatis.moteur.plugins.workflowsdb.WorkflowsDBException;
-import fr.insalyon.creatis.moteur.plugins.workflowsdb.dao.WorkflowsDBDAOException;
-import fr.insalyon.creatis.moteurlite.boutiques.BoutiquesService;
-import fr.insalyon.creatis.moteurlite.boutiques.scheme.BoutiquesDescriptor;
-import fr.insalyon.creatis.moteurlite.boutiques.scheme.Input;
-import fr.insalyon.creatis.moteurlite.boutiques.scheme.OutputFile;
-import fr.insalyon.creatis.moteurlite.iteration.IterationService;
+import fr.insalyon.creatis.moteurlite.runner.MoteurLiteRunner;
 
 public class MoteurLite {
-
-    public static final String RESULTS_DIRECTORY = "results-directory";
-    private static final Logger logger = Logger.getLogger(MoteurLite.class);
-
     public static void main(String[] args) throws MoteurLiteException {
         if (args.length != 3) {
             throw new IllegalArgumentException("Exactly 3 arguments are required: workflowId, boutiquesFilePath, inputsFilePath.");
-        }
-        String workflowId = args[0];
-        String boutiquesFilePath = args[1];
-        String inputsFilePath = args[2];
-
-        checkArguments(workflowId, boutiquesFilePath, inputsFilePath);
-
-        new MoteurLite().runWorkflow(workflowId, boutiquesFilePath, inputsFilePath);
-    }
-
-    private static void checkArguments(String workflowId, String boutiquesFilePath, String inputsFilePath) {
-        if (workflowId.trim().isEmpty()) {
+        } else if (args[0].trim().isEmpty()) {
             throw new IllegalArgumentException("Invalid workflowId. It should be a simple non-empty string.");
-        }
-
-        if ( ! boutiquesFilePath.endsWith(".json")) {
+        } else if ( ! args[1].endsWith(".json")) {
             throw new IllegalArgumentException("Invalid boutiquesFilePath. It should be a JSON file.");
-        }
-
-        if ( ! inputsFilePath.endsWith(".xml")) {
+        } else if ( ! args[2].endsWith(".xml")) {
             throw new IllegalArgumentException("Invalid inputsFilePath. It should be an XML file.");
+        } else {
+            new MoteurLiteRunner().run(args[1], args[2], args[3]);
         }
-    }
-
-    private final WorkflowsDbRepository workflowsDbRepo;
-    private final BoutiquesService boutiquesService;
-    private final InputsFileService inputsFileService;
-    private final IterationService iterationService;
-
-    public MoteurLite() throws MoteurLiteException {
-        boutiquesService = new BoutiquesService();
-        inputsFileService = new InputsFileService();
-        iterationService = new IterationService(boutiquesService);
-
-        try {
-            workflowsDbRepo = WorkflowsDbRepository.getInstance();
-        } catch (WorkflowsDBDAOException | WorkflowsDBException e) {
-            logger.error("Error creating workflows db repo", e);
-            throw new MoteurLiteException("Error creating workflows db repo", e);
-        }
-    }
-
-    public void runWorkflow(String workflowId, String boutiquesFilePath, String inputsFilePath) throws MoteurLiteException {
-        Map<String, List<String>> allInputs = inputsFileService.parseInputData(inputsFilePath);
-        BoutiquesDescriptor descriptor = boutiquesService.parseFile(boutiquesFilePath);
-        Map<String, Input> boutiquesInputs = boutiquesService.getInputsMap(descriptor);
-        HashMap<String, OutputFile> boutiquesOutputs = new HashMap<>(boutiquesService.getOutputMap(descriptor));
-
-        List<Map<String, String>> invocationsInputs = iterationService.compute(allInputs, descriptor);
-
-        workflowsDbRepo.persistProcessors(workflowId, descriptor.getName(), 0, 0, 0);
-        workflowsDbRepo.persistInputs(workflowId, allInputs, boutiquesInputs);
-
-        Gasw gasw = null;
-        try {
-            gasw = Gasw.getInstance();
-            GaswMonitor gaswMonitor = new GaswMonitor(gasw, workflowsDbRepo, workflowId, descriptor.getName(), boutiquesOutputs, invocationsInputs.size());
-            gasw.setNotificationClient(gaswMonitor);
-            gaswMonitor.start();
-        } catch (GaswException e) {
-            logger.error("Error launching gasw", e);
-            throw new MoteurLiteException("Error launching gasw", e);
-        }
-
-        createJobs(gasw, descriptor.getName(), invocationsInputs, boutiquesInputs);
-    }
-
-    private void createJobs(Gasw gasw, String applicationName, List<Map<String, String>> allInvocationsInputs, Map<String, Input> boutiquesInputs) throws MoteurLiteException {
-        for (Map<String, String> invocationInputs : allInvocationsInputs) {
-
-            URI resultsDirectoryURI = null;
-            List<URI> downloads = new ArrayList<>();
-            Map<String, String> finalInvocationInputs = new HashMap<>();
-
-            for (String inputId : invocationInputs.keySet()) {
-                String inputValue = invocationInputs.get(inputId);
-                if (RESULTS_DIRECTORY.equals(inputId)) {
-                    resultsDirectoryURI = getURI(inputValue);
-                } else {
-                    if (Input.Type.FILE.equals(boutiquesInputs.get(inputId).getType())) {
-                        URI downloadURI = getURI(inputValue);
-                        String filename = Paths.get(downloadURI.getPath()).getFileName().toString();
-                        downloads.add(downloadURI);
-                        inputValue = filename;
-                    }
-                    finalInvocationInputs.put(inputId, inputValue);
-                }
-            }
-
-            String invocationString = convertMapToJson(finalInvocationInputs, boutiquesInputs);
-            String jobId = applicationName + "-" + System.nanoTime() + ".sh";
-
-            GaswInput gaswInput = new GaswInput(applicationName, applicationName + ".json", downloads, resultsDirectoryURI, invocationString, jobId);
-            try {
-                gasw.submit(gaswInput);
-            } catch (GaswException e) {
-                logger.error("Error submitting gasw job", e);
-                throw new MoteurLiteException("Error submitting gasw job", e);
-            }
-        }
-    }
-
-    private URI getURI(String inputValue) throws MoteurLiteException {
-        try {
-            return new URI(inputValue);
-        } catch (URISyntaxException e) {
-            logger.error("Error parsing URI : " + inputValue, e);
-            throw new MoteurLiteException("Error parsing URI : " + inputValue, e);
-        }
-    }
-
-    private String convertMapToJson(Map<String, String> invocationInputs, Map<String, Input> boutiquesInputs) {
-        ObjectMapper mapper = new ObjectMapper();
-        ObjectNode jsonNode = mapper.createObjectNode();
-
-        for (String inputId : invocationInputs.keySet()) {
-            String value = invocationInputs.get(inputId);
-            Input.Type type = boutiquesInputs.get(inputId).getType();
-
-            if (type == Input.Type.NUMBER ) {
-                if (boutiquesInputs.get(inputId).getInteger() != null && boutiquesInputs.get(inputId).getInteger()) {
-                    jsonNode.put(inputId, Integer.parseInt(value));
-                } else {
-                    jsonNode.put(inputId, Float.parseFloat(value));
-                }
-            } else if (type == Input.Type.FLAG) {
-                jsonNode.put(inputId, Boolean.parseBoolean(value));
-            } else {
-                jsonNode.put(inputId, value);
-            }
-        }
-        return jsonNode.toString();
     }
 }
