+++ src/main/java/fr/insalyon/creatis/moteurlite/MoteurLiteConstants.java
@@ -1,0 +1,6 @@
+package fr.insalyon.creatis.moteurlite;
+
+public class MoteurLiteConstants {
+    
+    public static final String RESULTS_DIRECTORY = "results-directory";
+}
